+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/kinds/HistogramPanel/sections.ts
@@ -1,4 +1,4 @@
-import type { SectionConfig } from '../types';
+import type { SectionConfig } from '../../types/sections';
 
 export const sections: SectionConfig[] = [
 	{ kind: 'legend', controls: { position: true, mode: true } },
