+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/shared/baseConfigBuilder.ts
