+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/shared/getBuilderQueries.ts
