+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/shared/hooks/useTimeScale.ts
