+++ frontend/src/pages/DashboardPageV2/DashboardContainer/Panels/shared/panel.module.scss
