+++ frontend/src/pages/DashboardPageV2/DashboardContainer/PanelsAndSectionsLayout/Panel/hooks/usePanelInteractions.ts
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import { useLocation, useParams } from 'react-router-dom';
 import { QueryParams } from 'constants/query';
 import { PanelMode } from 'container/DashboardContainer/visualization/panels/types';
-import type { DashboardPreference } from 'pages/DashboardPageV2/DashboardContainer/Panels/types';
+import type { DashboardPreference } from 'pages/DashboardPageV2/DashboardContainer/Panels/types/rendererProps';
 import { useDashboardCursorSyncMode } from 'hooks/dashboard/useDashboardCursorSyncMode';
 import { useSyncTooltipFilterMode } from 'hooks/dashboard/useSyncTooltipFilterMode';
 import { useSafeNavigate } from 'hooks/useSafeNavigate';
